feature: BIM-31419 Typo in 'app-document-object-icon.const.ts' file has been fixed.
