+++ projects/icons/src/declarations/constants/icons/app-document-container-icon.const.ts
@@ -3,28 +3,29 @@ import { IconDefinition } from '../../interfaces/icon-definition.interface';
 export const appDocumentContainerIcon: IconDefinition = {
   name: 'app-document-container',
   data: `<svg xmlns="http://www.w3.org/2000/svg" width="14" height="14" viewBox="0 0 14 14" fill="none">
-<mask id="mask0_3676_28970" style="mask-type:alpha" maskUnits="userSpaceOnUse" x="0" y="0" width="14" height="14">
-<path d="M3.63469 3.90465C3.63469 3.75561 3.75551 3.6348 3.90454 3.6348C4.05358 3.6348 4.17439 3.75561 4.17439 3.90465C4.17439 4.05368 4.05358 4.1745 3.90454 4.1745C3.75551 4.1745 3.63469 4.05368 3.63469 3.90465Z" fill="currentColor"/>
-<path d="M3.90454 5.41684C3.75551 5.41684 3.63469 5.53766 3.63469 5.68669C3.63469 5.83573 3.75551 5.95655 3.90454 5.95655C4.05358 5.95655 4.17439 5.83573 4.17439 5.68669C4.17439 5.53766 4.05358 5.41684 3.90454 5.41684Z" fill="currentColor"/>
-<path d="M3.63469 7.40732C3.63469 7.25828 3.75551 7.13747 3.90454 7.13747C4.05358 7.13747 4.17439 7.25828 4.17439 7.40732C4.17439 7.55635 4.05358 7.67717 3.90454 7.67717C3.75551 7.67717 3.63469 7.55635 3.63469 7.40732Z" fill="currentColor"/>
-<path d="M3.90454 8.95337C3.75551 8.95337 3.63469 9.07419 3.63469 9.22322C3.63469 9.37226 3.75551 9.49307 3.90454 9.49307C4.05358 9.49307 4.17439 9.37226 4.17439 9.22322C4.17439 9.07419 4.05358 8.95337 3.90454 8.95337Z" fill="currentColor"/>
-<path d="M4.53845 4.87204C4.53845 4.72301 4.65927 4.60219 4.8083 4.60219C4.95734 4.60219 5.07815 4.72301 5.07815 4.87204C5.07815 5.02108 4.95734 5.14189 4.8083 5.14189C4.65927 5.14189 4.53845 5.02108 4.53845 4.87204Z" fill="currentColor"/>
-<path d="M6.59082 4.60219C6.44179 4.60219 6.32097 4.72301 6.32097 4.87204C6.32097 5.02108 6.44179 5.14189 6.59082 5.14189C6.73986 5.14189 6.86068 5.02108 6.86068 4.87204C6.86068 4.72301 6.73986 4.60219 6.59082 4.60219Z" fill="currentColor"/>
-<path d="M8.04414 4.87204C8.04414 4.72301 8.16495 4.60219 8.31399 4.60219C8.46302 4.60219 8.58384 4.72301 8.58384 4.87204C8.58384 5.02108 8.46302 5.14189 8.31399 5.14189C8.16495 5.14189 8.04414 5.02108 8.04414 4.87204Z" fill="currentColor"/>
-<path d="M6.59082 6.25392C6.44179 6.25392 6.32097 6.37474 6.32097 6.52377C6.32097 6.67281 6.44179 6.79363 6.59082 6.79363C6.73986 6.79363 6.86068 6.67281 6.86068 6.52377C6.86068 6.37474 6.73986 6.25392 6.59082 6.25392Z" fill="currentColor"/>
-<path d="M4.53845 6.59267C4.53845 6.44363 4.65927 6.32281 4.8083 6.32281C4.95734 6.32281 5.07815 6.44363 5.07815 6.59267C5.07815 6.7417 4.95734 6.86252 4.8083 6.86252C4.65927 6.86252 4.53845 6.7417 4.53845 6.59267Z" fill="currentColor"/>
-<path d="M4.8083 7.99826C4.65927 7.99826 4.53845 8.11908 4.53845 8.26811C4.53845 8.41715 4.65927 8.53797 4.8083 8.53797C4.95734 8.53797 5.07815 8.41715 5.07815 8.26811C5.07815 8.11908 4.95734 7.99826 4.8083 7.99826Z" fill="currentColor"/>
-<path d="M5.4295 3.90465C5.4295 3.75561 5.55031 3.6348 5.69935 3.6348C5.84838 3.6348 5.9692 3.75561 5.9692 3.90465C5.9692 4.05368 5.84838 4.1745 5.69935 4.1745C5.55031 4.1745 5.4295 4.05368 5.4295 3.90465Z" fill="currentColor"/>
-<path d="M7.48187 3.6348C7.33284 3.6348 7.21202 3.75561 7.21202 3.90465C7.21202 4.05368 7.33284 4.1745 7.48187 4.1745C7.63091 4.1745 7.75173 4.05368 7.75173 3.90465C7.75173 3.75561 7.63091 3.6348 7.48187 3.6348Z" fill="currentColor"/>
-<path d="M8.93518 3.90465C8.93518 3.75561 9.056 3.6348 9.20504 3.6348C9.35407 3.6348 9.47489 3.75561 9.47489 3.90465C9.47489 4.05368 9.35407 4.1745 9.20504 4.1745C9.056 4.1745 8.93518 4.05368 8.93518 3.90465Z" fill="currentColor"/>
-<path d="M5.69935 5.41684C5.55031 5.41684 5.4295 5.53766 5.4295 5.68669C5.4295 5.83573 5.55031 5.95655 5.69935 5.95655C5.84838 5.95655 5.9692 5.83573 5.9692 5.68669C5.9692 5.53766 5.84838 5.41684 5.69935 5.41684Z" fill="currentColor"/>
-<path d="M7.21202 5.68669C7.21202 5.53766 7.33284 5.41684 7.48187 5.41684C7.63091 5.41684 7.75173 5.53766 7.75173 5.68669C7.75173 5.83573 7.63091 5.95655 7.48187 5.95655C7.33284 5.95655 7.21202 5.83573 7.21202 5.68669Z" fill="currentColor"/>
-<path d="M5.69935 7.13747C5.55031 7.13747 5.4295 7.25828 5.4295 7.40732C5.4295 7.55635 5.55031 7.67717 5.69935 7.67717C5.84838 7.67717 5.9692 7.55635 5.9692 7.40732C5.9692 7.25828 5.84838 7.13747 5.69935 7.13747Z" fill="currentColor"/>
-<path fill-rule="evenodd" clip-rule="evenodd" d="M12.016 4.99377L12.0102 4.98799L13.2534 6.23121C13.5951 6.57292 13.5951 7.12694 13.2534 7.46865L7.46854 13.2535C7.12683 13.5952 6.57281 13.5952 6.2311 13.2535L5.22762 12.25H2.625C2.14175 12.25 1.75 11.8582 1.75 11.375V2.625C1.75 2.14175 2.14175 1.75 2.625 1.75H11.375C11.8582 1.75 12.25 2.14175 12.25 2.625V5.22778L12.016 4.99377ZM2.625 2.625H11.375V4.39035C11.039 4.18969 10.5974 4.23403 10.3081 4.52337L4.52327 10.3082C4.23396 10.5975 4.18959 11.039 4.39017 11.375H2.625L2.625 2.625ZM6.19413 9.87478L5.14198 10.9269L6.84982 12.6348L12.6347 6.84993L10.9268 5.14209L9.89386 6.17505C9.89902 6.17975 9.90411 6.1846 9.9091 6.18959L10.7682 7.04865C10.939 7.21951 10.939 7.49652 10.7682 7.66737C10.5973 7.83823 10.3203 7.83823 10.1494 7.66737L9.29038 6.80831C9.28539 6.80332 9.28054 6.79823 9.27584 6.79307L8.39454 7.67437C8.39971 7.67907 8.40479 7.68391 8.40978 7.6889L9.26885 8.54797C9.4397 8.71882 9.4397 8.99583 9.26885 9.16668C9.09799 9.33754 8.82098 9.33754 8.65013 9.16668L7.79107 8.30762C7.78607 8.30263 7.78123 8.29755 7.77653 8.29238L6.81214 9.25676C6.81732 9.26147 6.8224 9.26632 6.8274 9.27131L7.68646 10.1304C7.85732 10.3012 7.85732 10.5782 7.68646 10.7491C7.51561 10.92 7.2386 10.92 7.06774 10.7491L6.20868 9.89003C6.20368 9.88504 6.19883 9.87995 6.19413 9.87478Z" fill="currentColor"/>
-<path d="M9.34707 11.375L8.47207 12.25H11.375C11.8582 12.25 12.25 11.8582 12.25 11.375V8.47207L11.375 9.34707V11.375H9.34707Z" fill="currentColor"/>
-</mask>
-<g mask="url(#mask0_3676_28970)">
-<rect width="14" height="14" fill="currentColor"/>
-</g>
-</svg>`,
+  <mask id="mask0_3676_28970" style="mask-type:alpha" maskUnits="userSpaceOnUse" x="0" y="0" width="14" height="14">
+  <path d="M3.63469 3.90465C3.63469 3.75561 3.75551 3.6348 3.90454 3.6348C4.05358 3.6348 4.17439 3.75561 4.17439 3.90465C4.17439 4.05368 4.05358 4.1745 3.90454 4.1745C3.75551 4.1745 3.63469 4.05368 3.63469 3.90465Z" fill="currentColor"/>
+  <path d="M3.90454 5.41684C3.75551 5.41684 3.63469 5.53766 3.63469 5.68669C3.63469 5.83573 3.75551 5.95655 3.90454 5.95655C4.05358 5.95655 4.17439 5.83573 4.17439 5.68669C4.17439 5.53766 4.05358 5.41684 3.90454 5.41684Z" fill="currentColor"/>
+  <path d="M3.63469 7.40732C3.63469 7.25828 3.75551 7.13747 3.90454 7.13747C4.05358 7.13747 4.17439 7.25828 4.17439 7.40732C4.17439 7.55635 4.05358 7.67717 3.90454 7.67717C3.75551 7.67717 3.63469 7.55635 3.63469 7.40732Z" fill="currentColor"/>
+  <path d="M3.90454 8.95337C3.75551 8.95337 3.63469 9.07419 3.63469 9.22322C3.63469 9.37226 3.75551 9.49307 3.90454 9.49307C4.05358 9.49307 4.17439 9.37226 4.17439 9.22322C4.17439 9.07419 4.05358 8.95337 3.90454 8.95337Z" fill="currentColor"/>
+  <path d="M4.53845 4.87204C4.53845 4.72301 4.65927 4.60219 4.8083 4.60219C4.95734 4.60219 5.07815 4.72301 5.07815 4.87204C5.07815 5.02108 4.95734 5.14189 4.8083 5.14189C4.65927 5.14189 4.53845 5.02108 4.53845 4.87204Z" fill="currentColor"/>
+  <path d="M6.59082 4.60219C6.44179 4.60219 6.32097 4.72301 6.32097 4.87204C6.32097 5.02108 6.44179 5.14189 6.59082 5.14189C6.73986 5.14189 6.86068 5.02108 6.86068 4.87204C6.86068 4.72301 6.73986 4.60219 6.59082 4.60219Z" fill="currentColor"/>
+  <path d="M8.04414 4.87204C8.04414 4.72301 8.16495 4.60219 8.31399 4.60219C8.46302 4.60219 8.58384 4.72301 8.58384 4.87204C8.58384 5.02108 8.46302 5.14189 8.31399 5.14189C8.16495 5.14189 8.04414 5.02108 8.04414 4.87204Z" fill="currentColor"/>
+  <path d="M6.59082 6.25392C6.44179 6.25392 6.32097 6.37474 6.32097 6.52377C6.32097 6.67281 6.44179 6.79363 6.59082 6.79363C6.73986 6.79363 6.86068 6.67281 6.86068 6.52377C6.86068 6.37474 6.73986 6.25392 6.59082 6.25392Z" fill="currentColor"/>
+  <path d="M4.53845 6.59267C4.53845 6.44363 4.65927 6.32281 4.8083 6.32281C4.95734 6.32281 5.07815 6.44363 5.07815 6.59267C5.07815 6.7417 4.95734 6.86252 4.8083 6.86252C4.65927 6.86252 4.53845 6.7417 4.53845 6.59267Z" fill="currentColor"/>
+  <path d="M4.8083 7.99826C4.65927 7.99826 4.53845 8.11908 4.53845 8.26811C4.53845 8.41715 4.65927 8.53797 4.8083 8.53797C4.95734 8.53797 5.07815 8.41715 5.07815 8.26811C5.07815 8.11908 4.95734 7.99826 4.8083 7.99826Z" fill="currentColor"/>
+  <path d="M5.4295 3.90465C5.4295 3.75561 5.55031 3.6348 5.69935 3.6348C5.84838 3.6348 5.9692 3.75561 5.9692 3.90465C5.9692 4.05368 5.84838 4.1745 5.69935 4.1745C5.55031 4.1745 5.4295 4.05368 5.4295 3.90465Z" fill="currentColor"/>
+  <path d="M7.48187 3.6348C7.33284 3.6348 7.21202 3.75561 7.21202 3.90465C7.21202 4.05368 7.33284 4.1745 7.48187 4.1745C7.63091 4.1745 7.75173 4.05368 7.75173 3.90465C7.75173 3.75561 7.63091 3.6348 7.48187 3.6348Z" fill="currentColor"/>
+  <path d="M8.93518 3.90465C8.93518 3.75561 9.056 3.6348 9.20504 3.6348C9.35407 3.6348 9.47489 3.75561 9.47489 3.90465C9.47489 4.05368 9.35407 4.1745 9.20504 4.1745C9.056 4.1745 8.93518 4.05368 8.93518 3.90465Z" fill="currentColor"/>
+  <path d="M5.69935 5.41684C5.55031 5.41684 5.4295 5.53766 5.4295 5.68669C5.4295 5.83573 5.55031 5.95655 5.69935 5.95655C5.84838 5.95655 5.9692 5.83573 5.9692 5.68669C5.9692 5.53766 5.84838 5.41684 5.69935 5.41684Z" fill="currentColor"/>
+  <path d="M7.21202 5.68669C7.21202 5.53766 7.33284 5.41684 7.48187 5.41684C7.63091 5.41684 7.75173 5.53766 7.75173 5.68669C7.75173 5.83573 7.63091 5.95655 7.48187 5.95655C7.33284 5.95655 7.21202 5.83573 7.21202 5.68669Z" fill="currentColor"/>
+  <path d="M5.69935 7.13747C5.55031 7.13747 5.4295 7.25828 5.4295 7.40732C5.4295 7.55635 5.55031 7.67717 5.69935 7.67717C5.84838 7.67717 5.9692 7.55635 5.9692 7.40732C5.9692 7.25828 5.84838 7.13747 5.69935 7.13747Z" fill="currentColor"/>
+  <path fill-rule="evenodd" clip-rule="evenodd" d="M12.016 4.99377L12.0102 4.98799L13.2534 6.23121C13.5951 6.57292 13.5951 7.12694 13.2534 7.46865L7.46854 13.2535C7.12683 13.5952 6.57281 13.5952 6.2311 13.2535L5.22762 12.25H2.625C2.14175 12.25 1.75 11.8582 1.75 11.375V2.625C1.75 2.14175 2.14175 1.75 2.625 1.75H11.375C11.8582 1.75 12.25 2.14175 12.25 2.625V5.22778L12.016 4.99377ZM2.625 2.625H11.375V4.39035C11.039 4.18969 10.5974 4.23403 10.3081 4.52337L4.52327 10.3082C4.23396 10.5975 4.18959 11.039 4.39017 11.375H2.625L2.625 2.625ZM6.19413 9.87478L5.14198 10.9269L6.84982 12.6348L12.6347 6.84993L10.9268 5.14209L9.89386 6.17505C9.89902 6.17975 9.90411 6.1846 9.9091 6.18959L10.7682 7.04865C10.939 7.21951 10.939 7.49652 10.7682 7.66737C10.5973 7.83823 10.3203 7.83823 10.1494 7.66737L9.29038 6.80831C9.28539 6.80332 9.28054 6.79823 9.27584 6.79307L8.39454 7.67437C8.39971 7.67907 8.40479 7.68391 8.40978 7.6889L9.26885 8.54797C9.4397 8.71882 9.4397 8.99583 9.26885 9.16668C9.09799 9.33754 8.82098 9.33754 8.65013 9.16668L7.79107 8.30762C7.78607 8.30263 7.78123 8.29755 7.77653 8.29238L6.81214 9.25676C6.81732 9.26147 6.8224 9.26632 6.8274 9.27131L7.68646 10.1304C7.85732 10.3012 7.85732 10.5782 7.68646 10.7491C7.51561 10.92 7.2386 10.92 7.06774 10.7491L6.20868 9.89003C6.20368 9.88504 6.19883 9.87995 6.19413 9.87478Z" fill="currentColor"/>
+  <path d="M9.34707 11.375L8.47207 12.25H11.375C11.8582 12.25 12.25 11.8582 12.25 11.375V8.47207L11.375 9.34707V11.375H9.34707Z" fill="currentColor"/>
+  </mask>
+  <g mask="url(#mask0_3676_28970)">
+  <rect width="14" height="14" fill="currentColor"/>
+  </g>
+  </svg>
+  `,
 };
