+++ projects/icons/src/declarations/constants/icons/app-document-object-icon.const.ts
@@ -1,0 +1,14 @@
+import { IconDefinition } from '../../interfaces/icon-definition.interface';
+
+export const appDocumentObjectIcon: IconDefinition = {
+  name: 'app-document-object',
+  data: `<svg xmlns="http://www.w3.org/2000/svg" width="14" height="14" viewBox="0 0 14 14" fill="none">
+  <mask id="mask0_3676_29075" style="mask-type:alpha" maskUnits="userSpaceOnUse" x="0" y="0" width="14" height="14">
+  <path fill-rule="evenodd" clip-rule="evenodd" d="M1.75 3.35417C1.75 2.4734 2.4734 1.75 3.35417 1.75H10.6458C11.5266 1.75 12.25 2.4734 12.25 3.35417V10.6458C12.25 11.5266 11.5266 12.25 10.6458 12.25H3.35417C2.4734 12.25 1.75 11.5266 1.75 10.6458V3.35417ZM4.66962 2.625H3.35417C2.94635 2.625 2.625 2.94635 2.625 3.35417V4.66963L4.66962 2.625ZM5.90706 2.625L2.625 5.90706V7.61094L7.61093 2.625H5.90706ZM8.84837 2.625L2.625 8.84837V10.6443L10.6443 2.625H8.84837ZM11.3491 3.15767L3.15767 11.3491C3.22001 11.366 3.28588 11.375 3.35417 11.375H5.04649L11.375 5.04648V3.35417C11.375 3.28587 11.366 3.22001 11.3491 3.15767ZM11.375 6.28392L6.28392 11.375H7.98779L11.375 7.98779V6.28392ZM11.375 9.22523L9.22523 11.375H10.6458C11.0536 11.375 11.375 11.0536 11.375 10.6458V9.22523Z" fill="currentColor"/>
+  </mask>
+  <g mask="url(#mask0_3676_29075)">
+  <rect width="14" height="14" fill="currentColor"/>
+  </g>
+  </svg>
+  `,
+};
