+++ projects/icons/src/declarations/constants/icons/app-document-sheet-icon.const.ts
@@ -3,11 +3,12 @@ import { IconDefinition } from '../../interfaces/icon-definition.interface';
 export const appDocumentSheetIcon: IconDefinition = {
   name: 'app-document-sheet',
   data: `<svg xmlns="http://www.w3.org/2000/svg" width="14" height="14" viewBox="0 0 14 14" fill="none">
-<mask id="mask0_3676_29040" style="mask-type:alpha" maskUnits="userSpaceOnUse" x="0" y="0" width="14" height="14">
-<path fill-rule="evenodd" clip-rule="evenodd" d="M7.35982 0.875C7.59188 0.875 7.81444 0.967187 7.97854 1.13128L11.3378 4.49051C11.5019 4.6546 11.594 4.87716 11.594 5.10923V11.375C11.594 12.3415 10.8105 13.125 9.84404 13.125H3.71904C2.75255 13.125 1.96904 12.3415 1.96904 11.375V2.625C1.96904 1.6585 2.75255 0.875 3.71904 0.875H7.35982ZM7.10302 1.75H3.71904C3.23579 1.75 2.84404 2.14175 2.84404 2.625V11.375C2.84404 11.8582 3.23579 12.25 3.71904 12.25H9.84404C10.3273 12.25 10.719 11.8582 10.719 11.375V5.52098H8.41552C7.69065 5.52098 7.10302 4.93335 7.10302 4.20848V1.75ZM10.2558 4.64598L7.97802 2.3682V4.20848C7.97802 4.45011 8.1739 4.64598 8.41552 4.64598H10.2558Z" fill="currentColor"/>
-</mask>
-<g mask="url(#mask0_3676_29040)">
-<rect width="14" height="14" fill="currentColor"/>
-</g>
-</svg>`,
+  <mask id="mask0_3676_29040" style="mask-type:alpha" maskUnits="userSpaceOnUse" x="0" y="0" width="14" height="14">
+  <path fill-rule="evenodd" clip-rule="evenodd" d="M7.35982 0.875C7.59188 0.875 7.81444 0.967187 7.97854 1.13128L11.3378 4.49051C11.5019 4.6546 11.594 4.87716 11.594 5.10923V11.375C11.594 12.3415 10.8105 13.125 9.84404 13.125H3.71904C2.75255 13.125 1.96904 12.3415 1.96904 11.375V2.625C1.96904 1.6585 2.75255 0.875 3.71904 0.875H7.35982ZM7.10302 1.75H3.71904C3.23579 1.75 2.84404 2.14175 2.84404 2.625V11.375C2.84404 11.8582 3.23579 12.25 3.71904 12.25H9.84404C10.3273 12.25 10.719 11.8582 10.719 11.375V5.52098H8.41552C7.69065 5.52098 7.10302 4.93335 7.10302 4.20848V1.75ZM10.2558 4.64598L7.97802 2.3682V4.20848C7.97802 4.45011 8.1739 4.64598 8.41552 4.64598H10.2558Z" fill="currentColor"/>
+  </mask>
+  <g mask="url(#mask0_3676_29040)">
+  <rect width="14" height="14" fill="currentColor"/>
+  </g>
+  </svg>
+  `,
 };
