+++ projects/icons/src/declarations/constants/icons/app-download-import-icon.const.ts
@@ -3,6 +3,7 @@ import { IconDefinition } from '../../interfaces/icon-definition.interface';
 export const appDownloadImportIcon: IconDefinition = {
   name: 'app-download-import',
   data: `<svg xmlns="http://www.w3.org/2000/svg" width="12" height="12" viewBox="0 0 12 12" fill="none">
-<path d="M5.99373 8.9229C5.87988 8.92123 5.77117 8.87525 5.69067 8.79472L3.94067 7.04472C3.89869 7.00441 3.86516 6.95612 3.84207 6.90269C3.81898 6.84926 3.80678 6.79176 3.80619 6.73356C3.8056 6.67535 3.81663 6.61761 3.83863 6.56372C3.86063 6.50983 3.89316 6.46088 3.93432 6.41972C3.97548 6.37856 4.02444 6.34602 4.07833 6.32402C4.13222 6.30202 4.18995 6.29099 4.24816 6.29159C4.30636 6.29218 4.36386 6.30438 4.4173 6.32747C4.47073 6.35056 4.51901 6.38408 4.55933 6.42607L5.5625 7.42924V1.19373C5.56168 1.13576 5.57239 1.0782 5.59401 1.0244C5.61563 0.9706 5.64772 0.921636 5.68843 0.880349C5.72914 0.839063 5.77764 0.806278 5.83113 0.783901C5.88462 0.761524 5.94202 0.75 6 0.75C6.05798 0.75 6.11538 0.761524 6.16887 0.783901C6.22236 0.806278 6.27086 0.839063 6.31157 0.880349C6.35228 0.921636 6.38437 0.9706 6.40599 1.0244C6.42761 1.0782 6.43832 1.13576 6.4375 1.19373V7.42924L7.44067 6.42607C7.48099 6.38408 7.52927 6.35056 7.5827 6.32747C7.63614 6.30438 7.69364 6.29218 7.75184 6.29159C7.81005 6.29099 7.86778 6.30202 7.92167 6.32402C7.97556 6.34602 8.02452 6.37856 8.06568 6.41972C8.10684 6.46088 8.13937 6.50983 8.16137 6.56372C8.18338 6.61761 8.1944 6.67535 8.19381 6.73356C8.19322 6.79176 8.18102 6.84926 8.15793 6.90269C8.13484 6.95612 8.10131 7.00441 8.05933 7.04472L6.30933 8.79472C6.26796 8.8361 6.21872 8.86877 6.16451 8.89078C6.11031 8.9128 6.05224 8.92372 5.99373 8.9229ZM2.35417 3.95838C1.4734 3.95838 0.75 4.68178 0.75 5.56254V9.64588C0.75 10.5266 1.4734 11.25 2.35417 11.25H9.64583C10.5266 11.25 11.25 10.5266 11.25 9.64588V5.56254C11.25 4.68178 10.5266 3.95838 9.64583 3.95838H9.0625C9.00453 3.95756 8.94697 3.96827 8.89317 3.98989C8.83937 4.0115 8.79041 4.0436 8.74912 4.08431C8.70783 4.12501 8.67505 4.17352 8.65267 4.22701C8.63029 4.28049 8.61877 4.3379 8.61877 4.39588C8.61877 4.45386 8.63029 4.51126 8.65267 4.56475C8.67505 4.61823 8.70783 4.66674 8.74912 4.70745C8.79041 4.74815 8.83937 4.78025 8.89317 4.80187C8.94697 4.82349 9.00453 4.8342 9.0625 4.83338H9.64583C10.0536 4.83338 10.375 5.15473 10.375 5.56254V9.64588C10.375 10.0537 10.0536 10.375 9.64583 10.375H2.35417C1.94635 10.375 1.625 10.0537 1.625 9.64588V5.56254C1.625 5.15473 1.94635 4.83338 2.35417 4.83338H2.9375C2.99547 4.8342 3.05303 4.82349 3.10683 4.80187C3.16063 4.78025 3.2096 4.74815 3.25088 4.70745C3.29217 4.66674 3.32495 4.61823 3.34733 4.56475C3.36971 4.51126 3.38123 4.45386 3.38123 4.39588C3.38123 4.3379 3.36971 4.28049 3.34733 4.22701C3.32495 4.17352 3.29217 4.12501 3.25088 4.08431C3.2096 4.0436 3.16063 4.0115 3.10683 3.98989C3.05303 3.96827 2.99547 3.95756 2.9375 3.95838H2.35417Z" fill="currentColor"/>
-</svg>`,
+  <path d="M5.99373 8.9229C5.87988 8.92123 5.77117 8.87525 5.69067 8.79472L3.94067 7.04472C3.89869 7.00441 3.86516 6.95612 3.84207 6.90269C3.81898 6.84926 3.80678 6.79176 3.80619 6.73356C3.8056 6.67535 3.81663 6.61761 3.83863 6.56372C3.86063 6.50983 3.89316 6.46088 3.93432 6.41972C3.97548 6.37856 4.02444 6.34602 4.07833 6.32402C4.13222 6.30202 4.18995 6.29099 4.24816 6.29159C4.30636 6.29218 4.36386 6.30438 4.4173 6.32747C4.47073 6.35056 4.51901 6.38408 4.55933 6.42607L5.5625 7.42924V1.19373C5.56168 1.13576 5.57239 1.0782 5.59401 1.0244C5.61563 0.9706 5.64772 0.921636 5.68843 0.880349C5.72914 0.839063 5.77764 0.806278 5.83113 0.783901C5.88462 0.761524 5.94202 0.75 6 0.75C6.05798 0.75 6.11538 0.761524 6.16887 0.783901C6.22236 0.806278 6.27086 0.839063 6.31157 0.880349C6.35228 0.921636 6.38437 0.9706 6.40599 1.0244C6.42761 1.0782 6.43832 1.13576 6.4375 1.19373V7.42924L7.44067 6.42607C7.48099 6.38408 7.52927 6.35056 7.5827 6.32747C7.63614 6.30438 7.69364 6.29218 7.75184 6.29159C7.81005 6.29099 7.86778 6.30202 7.92167 6.32402C7.97556 6.34602 8.02452 6.37856 8.06568 6.41972C8.10684 6.46088 8.13937 6.50983 8.16137 6.56372C8.18338 6.61761 8.1944 6.67535 8.19381 6.73356C8.19322 6.79176 8.18102 6.84926 8.15793 6.90269C8.13484 6.95612 8.10131 7.00441 8.05933 7.04472L6.30933 8.79472C6.26796 8.8361 6.21872 8.86877 6.16451 8.89078C6.11031 8.9128 6.05224 8.92372 5.99373 8.9229ZM2.35417 3.95838C1.4734 3.95838 0.75 4.68178 0.75 5.56254V9.64588C0.75 10.5266 1.4734 11.25 2.35417 11.25H9.64583C10.5266 11.25 11.25 10.5266 11.25 9.64588V5.56254C11.25 4.68178 10.5266 3.95838 9.64583 3.95838H9.0625C9.00453 3.95756 8.94697 3.96827 8.89317 3.98989C8.83937 4.0115 8.79041 4.0436 8.74912 4.08431C8.70783 4.12501 8.67505 4.17352 8.65267 4.22701C8.63029 4.28049 8.61877 4.3379 8.61877 4.39588C8.61877 4.45386 8.63029 4.51126 8.65267 4.56475C8.67505 4.61823 8.70783 4.66674 8.74912 4.70745C8.79041 4.74815 8.83937 4.78025 8.89317 4.80187C8.94697 4.82349 9.00453 4.8342 9.0625 4.83338H9.64583C10.0536 4.83338 10.375 5.15473 10.375 5.56254V9.64588C10.375 10.0537 10.0536 10.375 9.64583 10.375H2.35417C1.94635 10.375 1.625 10.0537 1.625 9.64588V5.56254C1.625 5.15473 1.94635 4.83338 2.35417 4.83338H2.9375C2.99547 4.8342 3.05303 4.82349 3.10683 4.80187C3.16063 4.78025 3.2096 4.74815 3.25088 4.70745C3.29217 4.66674 3.32495 4.61823 3.34733 4.56475C3.36971 4.51126 3.38123 4.45386 3.38123 4.39588C3.38123 4.3379 3.36971 4.28049 3.34733 4.22701C3.32495 4.17352 3.29217 4.12501 3.25088 4.08431C3.2096 4.0436 3.16063 4.0115 3.10683 3.98989C3.05303 3.96827 2.99547 3.95756 2.9375 3.95838H2.35417Z" fill="currentColor"/>
+  </svg>
+  `,
 };
